+++ lib/table_sync/event_actions/data_wrapper/destroy.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+class TableSync::EventActions::DataWrapper::Destroy < TableSync::EventActions::DataWrapper::Base
+  def type
+    :destroy
+  end
+
+  def each
+    yield(event_data)
+  end
+
+  def destroy?
+    true
+  end
+
+  def update?
+    false
+  end
+end